feat(components): :sparkles: features related to last commit
