+++ components/ModalEditAddTask/ModalClose.styled.tsx
@@ -1,8 +1,8 @@
 import { DialogClose, DialogCloseProps } from '@radix-ui/react-dialog';
 import styled from 'styled-components';
 
-export interface ModalCloseProps extends DialogCloseProps {
-  isCreate?: boolean;
+export interface ModalCloseProps {
+  isPrimary?: boolean;
   p?: number;
   textColor?: string;
   fontWeight?: string;
@@ -10,8 +10,9 @@ export interface ModalCloseProps extends DialogCloseProps {
 export const ModalCloseStyled = styled(DialogClose)<ModalCloseProps>`
   padding: ${({ p }) => p}px;
   background-color: ${(props) =>
-    props.isCreate ? props.theme.primaryClrRed : 'transparent'};
+    props.isPrimary ? props.theme.primaryClrRed : 'transparent'};
   color: ${(props) => (props.textColor ? props.textColor : props.theme.mainText)};
   font-weight: ${(props) => props.fontWeight};
-  border-radius: ${(props) => (props.isCreate ? '8' : '0')}px;
+  border-radius: ${(props) => (props.isPrimary ? '8' : '0')}px;
+  text-transform: capitalize;
 `;
